A test showing DBD::Mock->begin_work, commit and rollback don't detect ->prepare failure
This is reported in RT ticket #66815

Basically any call to ->begin_work, ->commit, >rollback made when
their internal call to ->prepare returns an invalid statement handle
will not notice and proceed to use it anyway.

In particular, this happens when a DBD::Mock::Session is in effect,
and the session state list has been exhausted.  When this happens, the
method in question fails with an unhelpful error "Can't use an
undefined value as an ARRAY reference at ../lib/DBD/Mock.pm line
635", masking the real reason.